Change to Composition from Inheritance
Before I used PatrollingGuard class which is a derived class from
ThirdPersonCharacter. This is not flexible. Later, I may add an AI that
helps the player. In this case, setting PatrollingGuard as the parent
class of ThirdPersonCharacter won't be efficient.

Therefore, instead of setting PatrollingGuard (that contains the array
of patrol points) as the parent class , I created a class for Patrol
points, PatrolRoute, which is a UActorComponent. This class, PatrolRoute
can be used as a composition (componenet) of ThirdPersonCharacter.

This way, it will be more flexible.